requirements.txt
vrchatapi>=1.20.0
flask
pyinstaller
psutil
requests
Pillow
pywebview
pystray